Fix a bug and add Game of Life model
Fixed a bug where a WebSocket connection and associated goroutines
wouldn't be cleaned up immediately, due to pumpMessages waiting for
readPump to stop while readPump is blocked waiting for the next message
to arrive. The fix is to have pumpMessages not wait for readPump to
stop. readPump will unblock and stop when conn.Close is called.

Implemented Conway's Game of Life on the backend as a first step towards
converting this project into a "multiplayer" Game of Life. The backend
communicates with the frontend by sending down an initial game state
when the connection is set up, and then sending the difference between
game states on each tick.