Use functions instead of methods, update tests
- Refactor so that the operations for reading from, writing to, and
closing a connection are now functions instead of methods.
Methods had the drawback of requiring implementations to be in different
files, or requiring existing implementations to be modified to support
new tests in the same file. Using functions allows tests to be located
in the same file without being forced to share the implementations of
these operations.
- Fix a race condition in the send buffer overflow test, where the
model was receiving Ticks before it received the incoming Diff.
- Add test for when there is an error writing to the connection.
- Remove redundant tests