Handle page visibility changes
Disconnect/reconnect the websocket connection when the page becomes
hidden/visible. This may save resources, and also addresses a bug on
Safari for iOS where putting the page in the background by, e.g.,
turning the screen off and back on would cause the connection to be
closed. Notably, just checking for a closed connection via
WebSocket.readyState when the page becomes visible and reconnecting if
needed would not fix the bug, because readyState changes *after* the
visibilitychange event fires.