Fix bug occurring when submitting without having made any changes
Fix a bug where pressing the submit button without having filled any
overlay cells would cause the WebSocket connection to close. This was
happening because the client would send "undefined", which is not
considered a valid message by the server. The solution is to do nothing
when the submit button is pressed and there are no filled overlay cells.