Fix bug where submitted changes could be lost
This commit fixes a bug where the user could attempt to submit changes
while the WebSocket connection is not in the READY state, leading to
those changes being lost. This was most easily reproduced by drawing
something, switching tabs/windows, switching back and quickly pressing
Enter or the submit button.