Fix species input bug
This commit fixes a bug where navigating backward and then forward in
the browser would result in a mismatch between the species selected
according to the species input, and the actual species being drawn with.
The solution is to treat the value in the species input as the single
source of truth, rather than replicating it in the `species` variable.